requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flake8==7.0.0
Flask==3.0.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
